notification/requirements/requirements.txt
flask
sqlalchemy
nats-py